add support for Bring Your Own Cloud in Private Preview
Add fields to the Cluster resource and data source to support BYOC. Users can
specify a CustomerCloudAccount during cluster creation to direct CockroachDB
Cloud to create the cluster in an AWS account, GCP project, or Azure
subscription that they own.

To support this feature we also upgrade the Go SDK version to v6.6.0.